refactor: move MainWindow into separate module
